Improve blockchain validation, balances, and mining

+++ ledger/transaction.go
@@ -3,5 +3,5 @@ package ledger
 type Transaction struct {
 	Sender   string
 	Receiver string
-	Amount   float64
+	Amount   int
 }
